[#74] Remove `raiseError` accessor param.
If an accessor cannot process an input value, it should just throw an exception instead.

There is a trade-off to this approach, in that the stacktrace for the accessor-thrown error is not preserved (replaced by `_raiseError()`), but accessors tend to be fairly simple in implementation, so it shouldn' be too difficult to locate the source of an error.